+++ src/main/java/com/stakevault/betting/stats/domain/model/TenantSchemaNotFoundException.java
@@ -1,6 +1,6 @@
 package com.stakevault.betting.stats.domain.model;
 
-public class TenantSchemaNotFoundException extends RuntimeException implements LocalizedDomainException {
+public class TenantSchemaNotFoundException extends LocalizedRuntimeException {
 
 	public TenantSchemaNotFoundException(TenantSchemaName schema) {
 		super("tenant schema not provisioned: " + schema.value());